+++ crates/alkahest-renderer/src/camera/projection.rs
@@ -16,10 +16,12 @@ pub enum CameraProjection {
 }
 
 impl CameraProjection {
+    /// Field of view in degrees
     pub fn perspective(fov: f32, near: f32) -> Self {
         Self::Perspective { fov, near }
     }
 
+    /// Field of view in degrees
     pub fn perspective_bounded(fov: f32, near: f32, far: f32) -> Self {
         Self::PerspectiveBounded { fov, near, far }
     }